additional tests, prepping for 0.3.0 release
