+++ pkg/cmd/version.go
@@ -28,6 +28,6 @@ var versionCmd = &cobra.Command{
 	Use:   "version",
 	Short: "Print the version number of kube-capacity",
 	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("kube-capacity version 0.2.0")
+		fmt.Println("kube-capacity version 0.3.0")
 	},
 }
